+++ common/src/main/java/com/cha102/diyla/diyForum/DiyForumVO.java
@@ -1,66 +1,44 @@
-package com.cha102.diyla.diyForum;
+package com.cha102.diyla.commodityModel;
 
-public class DiyForumVO {
+import java.io.Serializable;
+import java.sql.Timestamp;
+import java.util.List;
+
+public class DiyForumVO implements java.io.Serializable {
     private Integer artiNo;
     private Integer memId;
     private Integer diyNo;
     private String artiCont;
-    private Integer diyGrade;
-
-    public DiyForumVO() {
-    }
-
-    public DiyForumVO(Integer artiNo, Integer memId, Integer diyNo, String artiCont, Integer diyGrade) {
-        this.artiNo = artiNo;
-        this.memId = memId;
-        this.diyNo = diyNo;
-        this.artiCont = artiCont;
-        this.diyGrade =  diyGrade;
-    }
-
-    @Override
-	public String toString() {
-		return "DiyForumVO [artiNo=" + artiNo + ", memId=" + memId + ", diyNo=" + diyNo + ", artiCont=" + artiCont
-				+ ", diyGrade=" + diyGrade + "]";
-	}
+    private Byte diyGra;
 
     public Integer getArtiNo() {
         return artiNo;
     }
-
     public void setArtiNo(Integer artiNo) {
         this.artiNo = artiNo;
     }
-
     public Integer getMemId() {
         return memId;
     }
-
     public void setMemId(Integer memId) {
         this.memId = memId;
     }
-
     public Integer getDiyNo() {
         return diyNo;
     }
-
     public void setDiyNo(Integer diyNo) {
         this.diyNo = diyNo;
     }
-
     public String getArtiCont() {
         return artiCont;
     }
-
     public void setArtiCont(String artiCont) {
         this.artiCont = artiCont;
     }
-
-    public Integer getDiyGrade() {
-        return diyGrade;
+    public Byte getDiyGra() {
+        return diyGra;
     }
-
-    public void setDiyGrade(Integer diyGrade) {
-        this.diyGrade = diyGrade;
+    public void setDiyGra(Byte diyGra) {
+        this.diyGra = diyGra;
     }
 }
